+++ shuup_tests/importer/test_admin.py
@@ -4,13 +4,16 @@
 #
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
+import json
 import mock
 import pytest
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.utils.translation import activate
+from urllib.parse import parse_qs, urlparse
 
-from shuup.core.models import Product
+from shuup.core.models import BackgroundTaskExecution, BackgroundTaskExecutionStatus, Product
 from shuup.default_importer.importers import ProductImporter
+from shuup.importer.admin_module.import_views import ImportDetailView, ImportListView
 from shuup.importer.utils.importer import ImportMode
 from shuup.testing.factories import (
     get_default_product_type,
@@ -18,20 +21,16 @@ from shuup.testing.factories import (
     get_default_shop,
     get_default_tax_class,
 )
+from shuup.testing.utils import apply_request_middleware
 from shuup.utils.django_compat import reverse
 from shuup_tests.utils import SmartClient
-
-try:
-    from urllib.parse import parse_qs, urlparse
-except ImportError:
-    from urlparse import parse_qs, urlparse
 
 
 def do_importing(sku, name, lang, shop, import_mode=ImportMode.CREATE_UPDATE, client=None):
     is_update = True
-    import_path = reverse("shuup_admin:importer.import")
+    import_path = reverse("shuup_admin:importer.import.new")
     process_path = reverse("shuup_admin:importer.import_process")
-    original_product = None
+
     if not client:
         is_update = False
         client = SmartClient()
@@ -89,7 +88,8 @@ def do_importing(sku, name, lang, shop, import_mode=ImportMode.CREATE_UPDATE, cl
     data = {}
     data["import_mode"] = import_mode.value
     process_submit_path = "%s?%s" % (process_path, query_string)
-    client.soup(process_submit_path, data=data, method="post")
+    response = client.post(process_submit_path, data=data)
+    assert response.status_code == 302
 
     assert Product.objects.count() == 1
     product = Product.objects.first()
@@ -131,16 +131,58 @@ def test_admin(rf, admin_user):
 
 
 @pytest.mark.django_db
+def test_admin_views(rf, admin_user):
+    # list imports
+    activate("en")
+    shop = get_default_shop()
+    shop.staff_members.add(admin_user)
+    get_default_tax_class()
+    get_default_product_type()
+    get_default_sales_unit()
+
+    # list imports, should be blank
+    list_view_func = ImportListView.as_view()
+    detail_view_func = ImportDetailView.as_view()
+    payload = {"jq": json.dumps({"perPage": 100, "page": 1}), "shop": shop.pk}
+    request = apply_request_middleware(rf.get("/", data=payload), user=admin_user)
+
+    response = list_view_func(request)
+    assert response.status_code == 200
+    assert len(json.loads(response.content)["items"]) == 0
+
+    # import a product
+    do_importing("123", "test", "en", shop)
+
+    # should contain 1 successfull import
+    response = list_view_func(request)
+    assert response.status_code == 200
+    data = json.loads(response.content)
+    assert len(data["items"]) == 1
+    assert data["items"][0]["importer"] == "Product Importer"
+    assert data["items"][0]["status"] == "Success"
+
+    # check the details of the import
+    request = apply_request_middleware(rf.get("/"), user=admin_user)
+    response = detail_view_func(request, pk=data["items"][0]["_id"])
+    assert response.status_code == 200
+    response.render()
+    template = response.content.decode("utf-8")
+    assert "Import Results" in template
+    assert "The following items were imported as new" in template
+    assert ">test</a>" in template
+    assert "Zero items were updated" in template
+
+
+@pytest.mark.django_db
 def test_invalid_files(rf, admin_user):
     lang = "en"
-    import_mode = ImportMode.CREATE_UPDATE
     sku = "123"
     name = "test"
 
-    import_path = reverse("shuup_admin:importer.import")
+    import_path = reverse("shuup_admin:importer.import.new")
     process_path = reverse("shuup_admin:importer.import_process")
-    tax_class = get_default_tax_class()
-    product_type = get_default_product_type()
+    get_default_tax_class()
+    get_default_product_type()
     get_default_sales_unit()
 
     activate("en")
@@ -159,8 +201,6 @@ def test_invalid_files(rf, admin_user):
     }
     response = client.post(import_path, data=data)
     assert response.status_code == 302
-    query_string = urlparse(response["location"]).query
-    query = parse_qs(query_string)
 
     data = {  # data with missing `n`
         "importer": "product_importer",
@@ -178,7 +218,7 @@ def test_invalid_file_type(rf, admin_user):
     sku = "123"
     name = "test"
 
-    import_path = reverse("shuup_admin:importer.import")
+    import_path = reverse("shuup_admin:importer.import.new")
     process_path = reverse("shuup_admin:importer.import_process")
     get_default_tax_class()
     get_default_product_type()
@@ -201,7 +241,6 @@ def test_invalid_file_type(rf, admin_user):
     response = client.post(import_path, data=data)
     assert response.status_code == 302
     query_string = urlparse(response["location"]).query
-    query = parse_qs(query_string)
 
     data = {  # data with missing `n`
         "importer": "product_importer",
@@ -249,10 +288,10 @@ def test_remap(rf, admin_user):
     sku = "123"
     name = "test"
 
-    import_path = reverse("shuup_admin:importer.import")
+    import_path = reverse("shuup_admin:importer.import.new")
     process_path = reverse("shuup_admin:importer.import_process")
-    tax_class = get_default_tax_class()
-    product_type = get_default_product_type()
+    get_default_tax_class()
+    get_default_product_type()
     get_default_sales_unit()
 
     activate("en")
@@ -311,7 +350,8 @@ def test_remap(rf, admin_user):
     data["import_mode"] = import_mode.value
     data["remap[gtiin]"] = "Product:gtin"  # map gtin
     process_submit_path = "%s?%s" % (process_path, query_string)
-    client.soup(process_submit_path, data=data, method="post")
+    response = client.post(process_submit_path, data=data)
+    assert response.status_code == 302
     assert Product.objects.count() == 1
     product = Product.objects.first()
     assert product.gtin == "111"
@@ -320,7 +360,7 @@ def test_remap(rf, admin_user):
 @pytest.mark.django_db
 def test_download_examples(rf, admin_user):
     shop = get_default_shop()
-    import_url = reverse("shuup_admin:importer.import")
+    import_url = reverse("shuup_admin:importer.import.new")
 
     shop.staff_members.add(admin_user)
     client = SmartClient()
@@ -367,47 +407,6 @@ def test_download_examples(rf, admin_user):
 
 
 @pytest.mark.django_db
-def test_import_error(admin_user):
-    shop = get_default_shop()
-    shop.staff_members.add(admin_user)
-    get_default_tax_class()
-    get_default_product_type()
-    get_default_sales_unit()
-
-    client = SmartClient()
-    client.login(username="admin", password="password")
-
-    import_path = reverse("shuup_admin:importer.import")
-    process_path = reverse("shuup_admin:importer.import_process")
-
-    client = SmartClient()
-    client.login(username="admin", password="password")
-
-    csv_content = str.encode("sku;name\n123;Teste")
-    data = {
-        "importer": "product_importer",
-        "shop": shop.pk,
-        "language": "en",
-        "file": SimpleUploadedFile("file.csv", csv_content, content_type="text/csv"),
-    }
-    response = client.post(import_path, data=data)
-    assert response.status_code == 302
-    query_string = urlparse(response["location"]).query
-    query = parse_qs(query_string)
-    data = {
-        "importer": "product_importer",
-        "shop": shop.pk,
-        "language": "en",
-        "n": query.get("n"),
-    }
-    data = {}
-    process_submit_path = "%s?%s" % (process_path, query_string)
-    response = client.post(process_submit_path, data=data)
-    assert response.status_code == 302
-    assert "Failed to import the file." == list(response.wsgi_request._messages)[0].message
-
-
-@pytest.mark.django_db
 def test_custom_file_transformer_import(admin_user):
     shop = get_default_shop()
     shop.staff_members.add(admin_user)
@@ -418,7 +417,7 @@ def test_custom_file_transformer_import(admin_user):
     client = SmartClient()
     client.login(username="admin", password="password")
 
-    import_path = reverse("shuup_admin:importer.import")
+    import_path = reverse("shuup_admin:importer.import.new")
     process_path = reverse("shuup_admin:importer.import_process")
 
     client = SmartClient()
@@ -445,7 +444,7 @@ def test_custom_file_transformer_import(admin_user):
         }
         response = client.post(process_submit_path, data=data)
         assert response.status_code == 302
-        assert (
-            "Error! Not implemented: `DataImporter` -> `transform_file()`."
-            in list(response.wsgi_request._messages)[0].message
-        )
+        assert "The import was queued!" in list(response.wsgi_request._messages)[0].message
+        task_execution = BackgroundTaskExecution.objects.last()
+        assert task_execution.status == BackgroundTaskExecutionStatus.ERROR
+        assert "Error! Not implemented: `DataImporter` -> `transform_file()`." in task_execution.error_log