+++ CHANGELOG.md
@@ -16,9 +16,16 @@ List all changes after the last release here (newer on top). Each change on a se
 ### Add
 
 - Core: Allow a supplier to have multiple modules
+- Core: add option to store tasks in the database to collect results
+
+### Changed
+
+- Importers: enable importers to run asynchronously
+- Importers: change the admin views to show the list of import processes
 
 ### Fixed
 
+- Admin: fix picotable overflow issue
 - Notify: always overrride the current notification data with the new one
 
 ## [2.9.2] - 2021-05-26